Refactor SetupPage by extracting logic to a custom hook and UI into smaller components
