+++ packages/web/src/app/setup/types.ts
@@ -1,0 +1,6 @@
+export type DatabaseItem = {
+	id: string;
+	title: string;
+	icon: string | null;
+	description: string;
+};
